Addition of logging controller and package refactoring
