+++ mFinance/src/edu/psu/ist412/authentication/Security.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package Login;
+package edu.psu.ist412.authentication;
 
 /**
  *
