+++ mFinance/src/edu/psu/ist412/transactions/Transaction.java
@@ -3,7 +3,7 @@
  * To change this template file, choose Tools | Templates
  * and open the template in the editor.
  */
-package edu.psu.ist412.accountTransaction;
+package edu.psu.ist412.transactions;
 
 /**
  *
